feat: finalizar CRUD com endpoint PUT e validar ProdutoRequestDTO

+++ src/main/java/com/guiverme/estoque/service/ProdutoService.java
@@ -39,6 +39,9 @@ public class ProdutoService {
     }
 
     public void deletar(Long id) {
+        if (!repository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Produto não encontrado com o ID: " + id);
+        }
         repository.deleteById(id);
     }
